Fix linting on function values example
